+++ taxsystem/views.py
@@ -22,12 +22,11 @@ from allianceauth.eveonline.models import EveCharacter, EveCorporationInfo
 from esi.decorators import token_required
 
 # AA TaxSystem
-from taxsystem import forms
+from taxsystem import forms, tasks
 from taxsystem.api.helpers import get_corporation, get_manage_permission
 from taxsystem.helpers.views import add_info_to_context
 from taxsystem.models.logs import AdminLogs, PaymentHistory
 from taxsystem.models.tax import Members, OwnerAudit, Payments, PaymentSystem
-from taxsystem.tasks import clear_all_etags, update_all_taxsytem, update_corporation
 
 logger = logging.getLogger(__name__)
 
@@ -55,10 +54,10 @@ def admin(request):
             force_refresh = True
         if request.POST.get("run_clear_etag"):
             messages.info(request, _("Queued Clear All ETags"))
-            clear_all_etags.apply_async(priority=1)
+            tasks.clear_all_etags.apply_async(priority=1)
         if request.POST.get("run_taxsystem_updates"):
             messages.info(request, _("Queued Update All Taxsystem"))
-            update_all_taxsytem.apply_async(
+            tasks.update_all_taxsytem.apply_async(
                 kwargs={"force_refresh": force_refresh}, priority=7
             )
     context = {
@@ -174,7 +173,7 @@ def add_corp(request, token):
             comment=_("Added to Tax System"),
         ).save()
 
-    update_corporation.apply_async(
+    tasks.update_corporation.apply_async(
         args=[owner.pk], kwargs={"force_refresh": True}, priority=6
     )
     msg = _("{corporation_name} successfully added/updated to Tax System").format(
@@ -205,6 +204,13 @@ def approve_payment(request: WSGIRequest, corporation_id: int, payment_pk: int):
                 reason = form.cleaned_data["accept_info"]
                 payment = Payments.objects.get(account__owner=corp, pk=payment_pk)
                 if payment.is_pending or payment.is_needs_approval:
+                    msg = _(
+                        "Payment ID: {pid} - Amount: {amount} - Name: {name} approved"
+                    ).format(
+                        pid=payment.pk,
+                        amount=intcomma(payment.amount),
+                        name=payment.name,
+                    )
                     payment.request_status = Payments.RequestStatus.APPROVED
                     payment.reviser = request.user.profile.main_character.character_name
                     payment.save()
@@ -250,6 +256,13 @@ def undo_payment(request: WSGIRequest, corporation_id: int, payment_pk: int):
                 reason = form.cleaned_data["undo_reason"]
                 payment = Payments.objects.get(account__owner=corp, pk=payment_pk)
                 if payment.is_approved or payment.is_rejected:
+                    msg = _(
+                        "Payment ID: {pid} - Amount: {amount} - Name: {name} undone"
+                    ).format(
+                        pid=payment.pk,
+                        amount=intcomma(payment.amount),
+                        name=payment.name,
+                    )
                     # Ensure that the payment is not rejected
                     if not payment.is_rejected:
                         account = PaymentSystem.objects.get(
@@ -304,7 +317,7 @@ def reject_payment(request: WSGIRequest, corporation_id: int, payment_pk: int):
                         owner=corp, user=payment.account.user
                     )
                     account.save()
-                    msg = _("Payment ID: %s - Amount %s - Name: %s rejected") % (
+                    msg = _("Payment ID: %s - Amount: %s - Name: %s rejected") % (
                         payment.pk,
                         intcomma(payment.amount),
                         payment.name,
@@ -328,7 +341,7 @@ def reject_payment(request: WSGIRequest, corporation_id: int, payment_pk: int):
 @login_required
 @permissions_required(["taxsystem.manage_own_corp", "taxsystem.manage_corps"])
 @require_POST
-def switch_user(request: WSGIRequest, corporation_id: int, user_pk: int):
+def switch_user(request: WSGIRequest, corporation_id: int, payment_system_pk: int):
     msg = _("Invalid Method")
     corp = get_corporation(request, corporation_id)
 
@@ -343,7 +356,9 @@ def switch_user(request: WSGIRequest, corporation_id: int, user_pk: int):
         with transaction.atomic():
             form = forms.TaxSwitchUserForm(data=request.POST)
             if form.is_valid():
-                payment_system = PaymentSystem.objects.get(owner=corp, pk=user_pk)
+                payment_system = PaymentSystem.objects.get(
+                    owner=corp, pk=payment_system_pk
+                )
                 if payment_system.is_active:
                     payment_system.status = PaymentSystem.Status.DEACTIVATED
                     msg = _("Payment System User: %s deactivated") % payment_system.name
@@ -366,6 +381,8 @@ def switch_user(request: WSGIRequest, corporation_id: int, user_pk: int):
     return JsonResponse(data={"success": False, "message": msg}, status=400, safe=False)
 
 
+@login_required
+@permissions_required(["taxsystem.manage_own_corp", "taxsystem.manage_corps"])
 @csrf_exempt
 def update_tax_amount(request: WSGIRequest, corporation_id: int):
     if request.method == "POST":
@@ -400,10 +417,12 @@ def update_tax_amount(request: WSGIRequest, corporation_id: int):
     return JsonResponse({"message": _("Invalid request method")}, status=405)
 
 
+@login_required
+@permissions_required(["taxsystem.manage_own_corp", "taxsystem.manage_corps"])
 @csrf_exempt
 def update_tax_period(request: WSGIRequest, corporation_id: int):
     if request.method == "POST":
-        value = float(request.POST.get("value"))
+        value = int(request.POST.get("value"))
         msg = _("Please enter a valid number")
         try:
             if value < 0:
@@ -437,7 +456,7 @@ def update_tax_period(request: WSGIRequest, corporation_id: int):
 @login_required
 @permissions_required(["taxsystem.manage_own_corp", "taxsystem.manage_corps"])
 @require_POST
-def delete_user(request: WSGIRequest, corporation_id: int, member_pk: int):
+def delete_member(request: WSGIRequest, corporation_id: int, member_pk: int):
     msg = _("Invalid Method")
     corp = get_corporation(request, corporation_id)